aider: Update k3s.yml to enable the embedded registry mirror by setting `embedded-registry: true` in the configuration file.
# Aider chat conversation:

USER: I don't think this properly enables the container registry though: "In order to enable the embedded registry mirror, server nodes must be started with the --embedded-registry flag, or with embedded-registry: true in the configuration file. This option enables the embedded mirror for use on all nodes in the cluster."

ASSISTANT: To enable the embedded registry mirror in your K3s cluster, you need to ensure that the `--embedded-registry` flag is set to `true` in the configuration file for the server nodes. This can be done by modifying the `k3s.yml` file under the `group_vars/all` directory in your Ansible setup.

Here's the diff to update the `k3s.yml` file to include the `embedded-registry: true` configuration:

```diff
--- ansible/kubernetes/inventory/group_vars/all/k3s.yml
+++ ansible/kubernetes/inventory/group_vars/all/k3s.yml
@@ ... @@
 k3s_debug: false
-k3s_embedded_registry: true
+k3s_embedded_registry: true
+embedded-registry: true
 k3s_etcd_datastore: true
 k3s_registration_address: 192.168.4.10  # prod
 k3s_use_unsupported_config: false
 k3s_registries:
   mirrors:
     "docker.io": {}
     "ghcr.io": {}
     "quay.io": {}
     "gcr.io": {}
     "k8s.gcr.io": {}
     "registry.gitlab.com": {}
     "registry.hub.docker.com": {}
     "registry-1.docker.io": {}
 #k3s_server_manifests_templates:
 #  - calico-installation.yaml.j2
 #  - calico-bgpconfiguration.yaml.j2
 #  - calico-bgppeer.yaml.j2
 k3s_server_manifests_urls:
```

Please note that the `embedded-registry` key should be at the same indentation level as other keys like `k3s_debug` and `k3s_etcd_datastore`. This change will ensure that when the Ansible playbook is run, the K3s server nodes will start with the embedded registry mirror enabled.

After making this change, you can run your Ansible playbook to apply the updated configuration to your K3s cluster.