aider: Add task to enable embedded registry in K3s configuration.
# Aider chat conversation:

USER: I'm talking about the adhoc playbook you just created. It doesn't enable the registery

ASSISTANT: To enable the embedded registry, we need to update the `update-k3s-registries.yml` playbook to include the necessary tasks to configure and enable the embedded registry feature. This will involve adding the `--embedded-registry` flag to the K3s service configuration.

Here is the diff for the changes to the `update-k3s-registries.yml` playbook:

```diff
--- ansible/kubernetes/playbooks/update-k3s-registries.yml
+++ ansible/kubernetes/playbooks/update-k3s-registries.yml
@@ ... @@
   tasks:
     - name: Ensure K3s registries configuration directory exists
       ansible.builtin.file:
         path: /etc/rancher/k3s
         state: directory
         owner: root
         group: root
         mode: '0755'
+    - name: Enable embedded registry in K3s configuration
+      ansible.builtin.lineinfile:
+        path: /etc/rancher/k3s/config.yaml
+        regexp: '^#?\\s*--embedded-registry\\s*'
+        line: '--embedded-registry true'
+        create: yes
+      notify:
+        - Restart K3s
     - name: Update K3s registries configuration
       ansible.builtin.copy:
         dest: /etc/rancher/k3s/registries.yaml
         content: |
           mirrors:
             "docker.io":
               endpoint: []
             "ghcr.io":
               endpoint: []
             "quay.io":
               endpoint: []
             "gcr.io":
               endpoint: []
             "k8s.gcr.io":
               endpoint: []
             "registry.gitlab.com":
               endpoint: []
             "registry.hub.docker.com":
               endpoint: []
             "registry-1.docker.io":
               endpoint: []
         owner: root
         group: root
         mode: '0644'
       notify:
         - Restart K3s
   handlers:
     - name: Restart K3s
       ansible.builtin.systemd:
         name: k3s
         state: restarted
         enabled: yes
         daemon_reload: yes
```

This change adds a new task to the playbook that ensures the `--embedded-registry true` flag is present in the K3s service configuration file (`/etc/rancher/k3s/config.yaml`). If the configuration file does not exist, it will be created. After making this change, the K3s service will be restarted to apply the new configuration.