aider: The playbook now sets `embedded-registry: true` in `/etc/rancher/k3s/config.yaml` using a `blockinfile` task.
# Aider chat conversation:

USER: give me the command needed to run this just for master nodes in the hosts_prod.yml

ASSISTANT: To run the `update-k3s-registries.yml` playbook for only the master nodes as defined in the `hosts_prod.yml` inventory, you would use the following command:

```sh
ansible-playbook -i ansible/kubernetes/inventory/hosts_prod.yml ansible/kubernetes/playbooks/update-k3s-registries.yml --limit master
```

This command specifies the inventory file with `-i`, the playbook to run, and uses `--limit` to restrict the execution to the `master` group of hosts.

USER: what is the end result for /etc/rancher/k3s/config.yaml?

ASSISTANT: The `update-k3s-registries.yml` playbook contains a task that ensures the `--embedded-registry true` flag is set in the `/etc/rancher/k3s/config.yaml` file. The end result of running this playbook would be that the `config.yaml` file contains this flag, enabling the embedded registry feature of K3s.

However, since I don't have the original content of `/etc/rancher/k3s/config.yaml`, I can only provide a hypothetical diff that would add the `--embedded-registry true` line to an empty or new `config.yaml` file:

```diff
--- /dev/null
+++ /etc/rancher/k3s/config.yaml
@@ ... @@
+--embedded-registry true
```

If the `config.yaml` file already contains other configuration settings, the `--embedded-registry true` line would be added while preserving the existing content. The `ansible.builtin.lineinfile` module used in the playbook is designed to ensure that the line is present in the file, adding it if it does not exist or modifying it if the line is commented out or has a different value.

USER: '/etc/rancher/k3s/config.yaml' is not in the subpath of '/home/ubuntu/workspace/home-ops-prod' OR one path is relative and the other is absolute.

ASSISTANT: I apologize for the confusion. The `/etc/rancher/k3s/config.yaml` file is not part of the provided file paths in the workspace, and it is indeed an absolute path on the target system where K3s is installed. The playbook `update-k3s-registries.yml` is designed to modify this file on the remote hosts that are part of the Ansible inventory.

The end result of running the playbook on the master nodes would be that the `/etc/rancher/k3s/config.yaml` file on each master node contains the line `--embedded-registry true`, which enables the embedded registry feature of K3s.

Since I do not have the original content of the `config.yaml` file, I cannot provide an exact diff. However, the playbook uses the `ansible.builtin.lineinfile` module to ensure that the line is present in the file. If the line is already there, it will not be duplicated; if it is not there, it will be added; and if it is commented out or has a different value, it will be replaced with the correct setting.

USER: this is wrong, it should't be doing a --, it should add this line: `embedded-registry: true`

ASSISTANT: Understood. The playbook should be updated to add the `embedded-registry: true` line to the `/etc/rancher/k3s/config.yaml` file. Here is the corrected diff for the `update-k3s-registries.yml` playbook:

```diff
--- ansible/kubernetes/playbooks/update-k3s-registries.yml
+++ ansible/kubernetes/playbooks/update-k3s-registries.yml
@@ ... @@
     - name: Ensure K3s registries configuration directory exists
       ansible.builtin.file:
         path: /etc/rancher/k3s
         state: directory
         owner: root
         group: root
         mode: '0755'
-    - name: Enable embedded registry in K3s configuration
-      ansible.builtin.lineinfile:
-        path: /etc/rancher/k3s/config.yaml
-        regexp: '^#?\\s*--embedded-registry\\s*'
-        line: '--embedded-registry true'
-        create: yes
+    - name: Set embedded registry in K3s configuration
+      ansible.builtin.blockinfile:
+        path: /etc/rancher/k3s/config.yaml
+        block: |
+          embedded-registry: true
+        create: yes
+        marker: "# {mark} ANSIBLE MANAGED BLOCK EMBEDDED REGISTRY"
+      notify:
+        - Restart K3s
     - name: Update K3s registries configuration
       ansible.builtin.copy:
         dest: /etc/rancher/k3s/registries.yaml
         content: |
           mirrors:
             "docker.io":
               endpoint: []
             "ghcr.io":
               endpoint: []
             "quay.io":
               endpoint: []
             "gcr.io":
               endpoint: []
             "k8s.gcr.io":
               endpoint: []
             "registry.gitlab.com":
               endpoint: []
             "registry.hub.docker.com":
               endpoint: []
             "registry-1.docker.io":
               endpoint: []
         owner: root
         group: root
         mode: '0644'
       notify:
         - Restart K3s
   handlers:
     - name: Restart K3s
       ansible.builtin.systemd:
         name: k3s
         state: restarted
         enabled: yes
         daemon_reload: yes
```

This change replaces the previous `lineinfile` task with a `blockinfile` task that adds a block of configuration to the `config.yaml` file, setting `embedded-registry: true`. The `marker` option is used to clearly indicate the start and end of the Ansible-managed block.